scanner/requirements.txt
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
google-oauth2-tool>=0.0.3
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
sentence-transformers>=2.2.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
duckdb>=0.8.0
pathlib>=1.0.1
torch>=1.9.0
transformers>=4.20.0
nltk>=3.8
spacy>=3.4.0
datasets>=2.0.0
networkx>=2.8
gensim>=4.2.0
textstat>=0.7.0
textblob>=0.17.0
vaderSentiment>=3.3.0
umap-learn>=0.5.0
scipy>=1.9.0